+++ record12.py
@@ -7,16 +7,24 @@ def display():
             print(f'{j}: {",".join(data[j])}')
 
 def create():
-    try:
-        f = open('Books.dat','rb')
-        f.close()
-    except FileNotFoundError:
-        f = open('Books.dat','wb')
+    def create_for(f):
         data = {}
         while True:
             data[input('Book No.: ')] = [input('Title: '),input('Price: ')]
             if input('Continue? (y/n): ') == 'n':break
         p.dump(data,f)
+    try:
+        f = open('Books.dat','rb')
+        f.close()
+        confirm = True if input('ARE YOU SURE YOU WANT TO OVERWRITE? (y/n): ') == 'y' else False
+        if confirm:
+            f = open('Books.dat','wb')
+            create_for(f)
+            f.close()
+        
+    except FileNotFoundError:
+        f = open('Books.dat','wb')
+        create_for(f)
         f.close()
 
 def update(bno):
@@ -30,4 +38,20 @@ def update(bno):
         else:
             print('BOOK NOT FOUND')
 
+while True:
+    print('''
+    1. CREATE
+    2. UPDATE
+    3. DISPLAY
+    4. EXIT
+    ''')
+    ch = int(input('ENTER YOUR CHOICE : '))
+    if ch == 1:
+        create()
+    elif ch == 2:
+        update(input('Book No.: '))
+    elif ch == 3:
+        display()
+    elif ch == 4:
+        break
